+++ rockpool/devices/xylo/syns65302/afe/__init__.py
@@ -1,5 +1,5 @@
 """
-Implements a simulation module for the digital audio front-end on Xylo-Audio 3
+Implements a simulation module for the digital audio front-end on XyloAudio 3
 
 Provides the modules :py:class:`.Raster`, :py:class:`.MicrophonePDM`, :py:class:`.PolyPhaseFIR`, :py:class:.`PDMADC`, :py:class:`.DivisiveNormalization`, :py:class:`.ChipButterworth`.
 """